refactor(broker): combine cluster services steps
