+++ broker/src/main/java/io/camunda/zeebe/broker/bootstrap/BrokerStartupProcess.java
@@ -53,7 +53,6 @@ public final class BrokerStartupProcess {
     result.add(new MonitoringServerStep());
     result.add(new BrokerAdminServiceStep());
 
-    result.add(new ClusterServicesCreationStep());
     result.add(new ClusterServicesStep());
 
     result.add(new CommandApiServiceStep());
